BUGFIX: Wrong normalization in filter resulting in instability
dt was being used for normalization on line 65 instead of dtsum. This
was causing the filer output to shoot up abnormally when temporarily
switching to another tab and coming back to the elm tab running the
filter after a while. The time gap results in a large dtsum, but dt is
limited to intervalLimit, so the normalization doesn't work out, blowing
up the value.