Changed Drag specification
Previously, the "Drag" force pinned the particle's position to the given
value. This is not desirable generally because it is often unclear which
point within the object is being pinned. Instead, the new specification
takes the change in the object's position, which is independent of the
point within the particle that's being pinned, and applies it to the
particle's position.