BREAKING CHANGE: Drag and drop approach
Changed the previous approach of treating the picker for a drag-n-drop
operation as an "Animation" type to a simple "Automaton" so that the
DnD behaviour is now controlled from the director instead of the
animator. The tasks are also generated by the director. I like this
much better.

The DnD direction seems to be now in the right place and the animation
code doesn't need to be called more often than requestAnimtaionFrame
just to keep pace with the director.

TODO: There is still quite some cruft in there - since the process of
computing deltas for picker movements should really be part of the
Anima module since it will be the same across applications.